improve move serializer to common
