+++ seatunnel-common/src/main/java/org/apache/seatunnel/common/serializer/ProtoStuffSerializer.java
@@ -18,7 +18,7 @@
  *
  */
 
-package org.apache.seatunnel.engine.imap.storage.api.common;
+package org.apache.seatunnel.common.serializer;
 
 import io.protostuff.LinkedBuffer;
 import io.protostuff.ProtostuffIOUtil;
