+++ seatunnel-engine/seatunnel-engine-storage/imap-storage-api/pom.xml
@@ -31,15 +31,7 @@
     <artifactId>imap-storage-api</artifactId>
 
     <dependencies>
-        <dependency>
-            <groupId>io.protostuff</groupId>
-            <artifactId>protostuff-core</artifactId>
-        </dependency>
 
-        <dependency>
-            <groupId>io.protostuff</groupId>
-            <artifactId>protostuff-runtime</artifactId>
-        </dependency>
     </dependencies>
 
 </project>
