+++ seatunnel-engine/seatunnel-engine-storage/imap-storage-file/pom.xml
@@ -29,8 +29,13 @@
     <modelVersion>4.0.0</modelVersion>
 
     <artifactId>imap-storage-file</artifactId>
-    
+
     <dependencies>
+        <dependency>
+            <groupId>org.apache.seatunnel</groupId>
+            <artifactId>seatunnel-common</artifactId>
+            <version>${project.version}</version>
+        </dependency>
         <dependency>
             <groupId>org.apache.seatunnel</groupId>
             <artifactId>imap-storage-api</artifactId>
