+++ seatunnel-engine/seatunnel-engine-storage/imap-storage-file/src/test/java/org/apache/seatunnel/engine/imap/storage/file/common/WALReaderAndWriterTest.java
@@ -24,8 +24,8 @@ import static org.awaitility.Awaitility.await;
 import static org.junit.jupiter.api.condition.OS.LINUX;
 import static org.junit.jupiter.api.condition.OS.MAC;
 
-import org.apache.seatunnel.engine.imap.storage.api.common.ProtoStuffSerializer;
-import org.apache.seatunnel.engine.imap.storage.api.common.Serializer;
+import org.apache.seatunnel.common.serializer.ProtoStuffSerializer;
+import org.apache.seatunnel.common.serializer.Serializer;
 import org.apache.seatunnel.engine.imap.storage.file.bean.IMapFileData;
 
 import org.apache.hadoop.conf.Configuration;
